pyproject.toml
[tool.poetry]
name = "kusto-db-explorer"
version = "0.1.0"
description = "A tool to explore Kusto databases and tables in a tree-like format"
authors = ["Contoso Developer"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.13"
azure-kusto-data = "^4.6.3"
azure-identity = "^1.20.0"
colorama = "^0.4.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
